下载完毕之后,配置文件写入到本地

+++ GameClient/Framework/Assets/ThirdPartyLibraries/Assets/HotUpdate/DownloadAssets.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
+using System.Text;
 using BestHTTP;
 using Common;
 using LitJson;
@@ -18,13 +19,14 @@ namespace GameAssets
     /// </summary>
     public class DownloadAssets : IBusiness
     {
-        protected struct DownloadABInfo
+        protected struct DownloadFileInfo
         {
-            public string abName;
+            public string fileName;//存入的有 AB 包的名字,也有压缩包的名字
             public bool downloadFinished;
+            public File_V_MD5 remoteFileVMd5;//记录远程的 File_V_MD5,当下载一个完毕之后,赋值给本地的版本配置文件对象
         }
         
-        private static Queue<DownloadABInfo> downloadQueue = new Queue<DownloadABInfo>(300);
+        private static Queue<DownloadFileInfo> downloadQueue = new Queue<DownloadFileInfo>(300);
 
         private static FileStream downloadFileStream;
         
@@ -35,7 +37,9 @@ namespace GameAssets
             Progress = 1;
             yield return DownloadVersionConfig();
             Progress = 15;
-            yield return DownloadAB();
+            yield return DownloadFiles();
+            Progress = 95;
+            yield return WriteToLocal();
             Progress = 100;
         }
         
@@ -87,37 +91,45 @@ namespace GameAssets
             {
                 yield break;//下载配置文件自身的平台,版本号,游戏二进制号 与本身的平台不匹配,不大于的情况下不给下载
             }
-            //查找配置表中的所有不存在.MD5 值不匹配的 ab 包,然后记录下来
-            foreach (var remoteItem in remoteVersionConfig.ABInfos)
+            else//有更新时,先将本地配置文件的版本数据重置
             {
-                var localItem = AssetsConfig.VersionConfig.ABInfos;
-                if (!localItem.TryGetValue(remoteItem.Key,out AB_V_MD5 localABVMd5))
+                AssetsConfig.VersionConfig.SVNVersion = remoteVersionConfig.SVNVersion;
+                AssetsConfig.VersionConfig.AppVersion = remoteVersionConfig.AppVersion;
+            }
+            
+            //查找配置表中的所有不存在与 MD5 值不匹配的 ab/zip 文件,然后记录下来
+            foreach (var remoteItem in remoteVersionConfig.FileInfos)
+            {
+                var localItem = AssetsConfig.VersionConfig.FileInfos;
+                if (!localItem.TryGetValue(remoteItem.Key,out File_V_MD5 localABVMd5))
                 {
-                    downloadQueue.Enqueue(new DownloadABInfo()
+                    downloadQueue.Enqueue(new DownloadFileInfo()
                     {
-                        abName = remoteItem.Key,
+                        fileName = remoteItem.Key,
                         downloadFinished = false,
+                        remoteFileVMd5 = remoteItem.Value,
                     });//本地没查到这个 AB,需要添加进下载队列
                 }
                 //本地和远程都查找到了,但是远程的版本大于本地的版本,并且 MD5 值不同,需要添加进下载队列
                 if (remoteItem.Value.Version.ToInt() > localABVMd5.Version.ToInt() && 
                     !remoteItem.Value.Md5Hash.Equals(localABVMd5.Md5Hash))
                 {
-                    downloadQueue.Enqueue(new DownloadABInfo()
+                    downloadQueue.Enqueue(new DownloadFileInfo()
                     {
-                        abName = remoteItem.Key,
+                        fileName = remoteItem.Key,
                         downloadFinished = false,
+                        remoteFileVMd5 = remoteItem.Value,
                     });
                 }
             }
             yield return AssetsConfig.OneFrame;
         }
-        
+
         /// <summary>
         /// 通过下载队列下载 AB 包
         /// </summary>
         /// <returns></returns>
-        public IEnumerator DownloadAB()
+        public IEnumerator DownloadFiles()
         {
             if (downloadQueue.Count <= 0)yield break;//没有下载列表
             
@@ -126,13 +138,14 @@ namespace GameAssets
             
             while (downloadQueue.Count <= 0)//循环下载队列,下载一个移除一个
             {
-                Progress = Progress + (int)(downloadQueue.Count/85);
-                DownloadABInfo info = downloadQueue.Peek();
-                string url = AssetsConfig.QueryRemoteABURL(info.abName);
-                string path = AssetsConfig.QueryDownloadABPath(info.abName);
-                File.Delete(path);//在这一步先进行删除,然后再下载
+                yield return AssetsConfig.OneFrame;
+                Progress = Progress + (int)(downloadQueue.Count/80);
+                DownloadFileInfo info = downloadQueue.Peek();
+                string url = AssetsConfig.QueryRemoteABURL(info.fileName);
+                string path = AssetsConfig.QueryDownloadABPath(info.fileName);
+                if (File.Exists(path)) File.Delete(path);//在这一步先进行删除,然后再下载
                 downloadFileStream = new FileStream(path, FileMode.OpenOrCreate, FileAccess.Write);
-                BestHttpHelper.Download(url,DownloadAB);
+                BestHttpHelper.Download(url,DownloadFiles);
                 while (!info.downloadFinished)
                 {
                     yield return AssetsConfig.OneFrame;
@@ -144,29 +157,44 @@ namespace GameAssets
             yield return AssetsConfig.OneFrame;
         }
         
-        private void DownloadAB(bool isError, byte[] data, int dataLength)
+        private void DownloadFiles(bool isError, byte[] data, int dataLength)
         {
             if (-100 == dataLength && null == data)
             {
-                DownloadABInfo info = downloadQueue.Peek();
                 //下载完毕,并且正常
-                downloadQueue.Dequeue();
                 downloadFileStream.Flush();
                 downloadFileStream.Close();
+                //下载正常之后,需要对这个包进行 MD5 校验
+                DownloadFileInfo info = downloadQueue.Dequeue();
+                string path = AssetsConfig.QueryDownloadABPath(info.fileName);
+                string localMD5 = SecurityTools.GetMD5Hash(path);
+                if (SecurityTools.VerifyMd5Hash(localMD5,info.remoteFileVMd5.Md5Hash))
+                {
+                    AssetsConfig.VersionConfig.FileInfos[info.fileName] = info.remoteFileVMd5;
+                }
+                else
+                {
+                    Debug.LogError("下载到本地的 MD5 与网络上面的 MD5 不匹配:" + 
+                                   info.fileName +"\n" + 
+                                   info.remoteFileVMd5 + "\n" + 
+                                   localMD5);
+                    AssetsNotification.Broadcast(IAssetsNotificationType.DownloadABSucceed,
+                        info.fileName + "的MD5 值,下载之后,本地与资源服务器上面的不匹配");
+                }
                 info.downloadFinished = true;
             }
             else if (isError && -200 == dataLength && null == data)
             {
-                DownloadABInfo info = downloadQueue.Peek();
+                DownloadFileInfo info = downloadQueue.Peek();
                 //下载过程中报错,需要删除之后,再重新添加进下载队列里面,再次重新下载
                 downloadQueue.Dequeue();
                 downloadQueue.Enqueue(info);
                 downloadFileStream.Flush();
                 downloadFileStream.Close();
-                File.Delete(AssetsConfig.QueryDownloadABPath(info.abName));
+                File.Delete(AssetsConfig.QueryDownloadABPath(info.fileName));
                 
                 AssetsNotification.Broadcast(IAssetsNotificationType.DownloadABFailed,
-                    "下载 AB 包失败一次" + info.abName);
+                    "下载 AB 包失败一次" + info.fileName);
                 
                 info.downloadFinished = true;
             }
@@ -179,8 +207,25 @@ namespace GameAssets
                 }
             }
         }
-        
-        
+
+        /// <summary>
+        /// 将更新过后的配置所有数据写入到本地
+        /// </summary>
+        /// <returns></returns>
+        private IEnumerator WriteToLocal()
+        {
+            //将所有数据,写入文件中.
+            using (FileStream fileStream = new FileStream(AssetsConfig.VersionConfigPersistentDataPath,
+                FileMode.Open,FileAccess.Write))
+            {
+                fileStream.SetLength(0);
+                fileStream.Flush();
+                byte[] data = Encoding.UTF8.GetBytes(JsonMapper.ToJson(AssetsConfig.VersionConfig));
+                fileStream.Write(data,0,data.Length);
+                fileStream.Flush();
+            }
+            yield return AssetsConfig.OneFrame;
+        }
 
     }
 }
